kube-spawn: renamed start to up
For better user experience, let's rename "start" command to "up".
